# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HelpHub Smart Service Desk API
  version: 1.0.0
  description: Secure employee and service-desk ticket workflows.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/tickets/{ticketId}/attachments:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      summary: List ticket attachments visible to the caller
      responses:
        '200':
          description: Attachment metadata ordered by upload time
    post:
      summary: Upload a ticket attachment
      description: Accepts PDF, PNG, JPEG, or plain text up to 5 MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Attachment stored
        '400':
          description: File type, signature, or size was rejected
  /api/v1/tickets/{ticketId}/attachments/{attachmentId}/content:
    parameters:
      - $ref: '#/components/parameters/TicketId'
      - name: attachmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Download an authorized ticket attachment
      responses:
        '200':
          description: Attachment bytes
        '404':
          description: Ticket or attachment is unavailable to the caller
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...